+++ tests/performance_tests/performance_test_basic.py
@@ -2,89 +2,20 @@
 
 import os
 import sys
-import re
 
 harnessPath = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(harnessPath)
 
-from testUtils import Account
 from testUtils import Utils
 from Cluster import Cluster
 from WalletMgr import WalletMgr
-from Node import Node
-from Node import ReturnType
 from TestHelper import TestHelper
-from dataclasses import dataclass
+import log_reader
 
 Print = Utils.Print
 errorExit = Utils.errorExit
 cmdError = Utils.cmdError
 relaunchTimeout = 30
-emptyBlockGoal = 5
-
-@dataclass
-class blockData():
-    partialBlockId: str = ""
-    blockNum: int = 0
-    transactions: int = 0
-    net: int = 0
-    cpu: int = 0
-    elapsed: int = 0
-    time: int = 0
-    latency: int = 0
-
-class chainData():
-    def __init__(self):
-        self.blockLog = []
-        self.startBlock = 0
-        self.ceaseBlock = 0
-        self.totalTransactions = 0
-        self.totalNet = 0
-        self.totalCpu = 0
-        self.totalElapsed = 0
-        self.totalTime = 0
-        self.totalLatency = 0
-    def updateTotal(self, transactions, net, cpu, elapsed, time, latency):
-        self.totalTransactions += transactions
-        self.totalNet += net
-        self.totalCpu += cpu
-        self.totalElapsed += elapsed
-        self.totalTime += time
-        self.totalLatency += latency
-    def __str__(self):
-        return (f"Starting block: {self.startBlock}\nEnding block:{self.ceaseBlock}\nChain transactions: {self.totalTransactions}\n"
-         f"Chain cpu: {self.totalNet}\nChain net: {self.totalCpu}\nChain elapsed: {self.totalElapsed}\nChain time: {self.totalTime}\nChain latency: {self.totalLatency}")
-
-def waitForEmptyBlocks(node):
-    emptyBlocks = 0
-    while emptyBlocks < emptyBlockGoal:
-        headBlock = node.getHeadBlockNum()
-        block = node.processCurlCmd("chain", "get_block_info", f'{{"block_num":{headBlock}}}', silentErrors=False, exitOnError=True)
-        node.waitForHeadToAdvance()
-        if block['transaction_mroot'] == "0000000000000000000000000000000000000000000000000000000000000000":
-            emptyBlocks += 1
-        else:
-            emptyBlocks = 0
-    return node.getHeadBlockNum()
-
-def fetchStats(total):
-    with open("var/lib/node_01/stderr.txt") as f:
-        blockResult = re.findall(r'Received block ([0-9a-fA-F]*).* #(\d+) .*trxs: (\d+)(.*)', f.read())
-        for value in blockResult:
-            v3Logging = re.findall(r'net: (\d+), cpu: (\d+), elapsed: (\d+), time: (\d+), latency: (-?\d+) ms', value[3])
-            if v3Logging:
-                total.blockLog.append(blockData(value[0], int(value[1]), int(value[2]), int(v3Logging[0][0]), int(v3Logging[0][1]), int(v3Logging[0][2]), int(v3Logging[0][3]), int(v3Logging[0][4])))
-                if int(value[1]) in range(total.startBlock, total.ceaseBlock):
-                    total.updateTotal(int(value[2]), int(v3Logging[0][0]), int(v3Logging[0][1]), int(v3Logging[0][2]), int(v3Logging[0][3]), int(v3Logging[0][4]))
-            else:
-                v2Logging = re.findall(r'latency: (-?\d+) ms', value[3])
-                if v2Logging:
-                    total.blockLog.append(blockData(value[0], int(value[1]), int(value[2]), 0, 0, 0, 0, int(v2Logging[0])))
-                    if int(value[1]) in range(total.startBlock, total.ceaseBlock):
-                        total.updateTotal(int(value[2]), 0, 0, 0, 0, int(v2Logging[0]))
-                else:
-                    print("Error: Unknown log format")
-
 
 args=TestHelper.parse_args({"-p","-n","-d","-s","--nodes-file"
                             ,"--dump-error-details","-v","--leave-running"
@@ -141,9 +72,9 @@ try:
     testGenerationDurationSec = 60
     targetTps = 1
     transactionsSent = testGenerationDurationSec * targetTps
-    data = chainData()
+    data = log_reader.chainData()
 
-    data.startBlock = waitForEmptyBlocks(validationNode)
+    data.startBlock = log_reader.waitForEmptyBlocks(validationNode)
 
     if Utils.Debug: Print(
                             f'Running trx_generator: ./tests/trx_generator/trx_generator  '
@@ -166,8 +97,8 @@ try:
                             f'--target-tps {targetTps}'
                          )
     # Get stats after transaction generation stops
-    data.ceaseBlock = waitForEmptyBlocks(validationNode) - emptyBlockGoal
-    fetchStats(data)
+    data.ceaseBlock = log_reader.waitForEmptyBlocks(validationNode) - log_reader.emptyBlockGoal + 1
+    log_reader.fetchStats(data, "var/lib/node_01/stderr.txt")
 
     print(data)
     assert transactionsSent == data.totalTransactions , f"Error: Transactions received: {data.totalTransactions} did not match expected total: {transactionsSent}"
