+++ tests/performance_tests/read_log_data.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+ 
+import argparse
+import log_reader
+
+parser = argparse.ArgumentParser(add_help=False)
+parser.add_argument("log_path", type=str, help="Path to nodeos log to scrape")
+parser.add_argument("--start_block", type=int, help="First significant block number in the log", default=2)
+args = parser.parse_args()
+logPath=args.log_path
+data = log_reader.chainData()
+data.startBlock = args.start_block
+log_reader.fetchStats(data, logPath)
+print(data)
+data.printBlockData()
